Update latest major version checking
